feat: conflicting control attributes diagnostic
