+++ src/M31.FluentApi.Generator/AnalyzerReleases.Unshipped.md
@@ -1,0 +1,5 @@
+### New Rules
+
+Rule ID | Category | Severity | Notes
+--------|----------|----------|-------
+M31FA019 | M31.Usage | Error | Conflicting control attributes
