+++ src/M31.FluentApi.Generator/SourceGenerators/Step.cs
@@ -1,0 +1,11 @@
+namespace M31.FluentApi.Generator.SourceGenerators;
+
+internal record Step
+{
+    public Step(int? value)
+    {
+        Value = value;
+    }
+
+    internal int? Value { get; }
+}
